[TASK] Split the functional tests for `DefaultController`
The `DefaultController` class is a good class with far too many
responsibilities. This also has the consequence that the related
test cases become far too big.

Split them up to make them more manageable, and to avoid having to
set up the world for each and every test.
