+++ Tests/Functional/FrontEnd/DefaultController/Fixtures/EventListWithCategories.xml
