+++ Tests/Functional/FrontEnd/DefaultController/Fixtures/EventListWithPlaces.xml
