+++ Tests/Functional/FrontEnd/DefaultController/Fixtures/EventListWithTypes.xml
