+++ Tests/Functional/FrontEnd/DefaultController/ListViewTest.php
@@ -2,12 +2,11 @@
 
 declare(strict_types=1);
 
-namespace OliverKlee\Seminars\Tests\Functional\FrontEnd;
+namespace OliverKlee\Seminars\Tests\Functional\FrontEnd\DefaultController;
 
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
 use OliverKlee\Oelib\System\Typo3Version;
 use OliverKlee\Oelib\Testing\CacheNullifyer;
-use OliverKlee\Seminars\FrontEnd\DefaultController;
 use OliverKlee\Seminars\Service\RegistrationManager;
 use OliverKlee\Seminars\Tests\Functional\Traits\LanguageHelper;
 use OliverKlee\Seminars\Tests\LegacyUnit\FrontEnd\Fixtures\TestingDefaultController;
@@ -24,7 +23,7 @@ use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
 /**
  * @covers \OliverKlee\Seminars\FrontEnd\DefaultController
  */
-final class DefaultControllerTest extends FunctionalTestCase
+final class ListViewTest extends FunctionalTestCase
 {
     use LanguageHelper;
 
@@ -97,72 +96,6 @@ final class DefaultControllerTest extends FunctionalTestCase
         $GLOBALS['TSFE'] = $frontEndController;
 
         return $frontEndController;
-    }
-
-    // Tests concerning the plugin definition
-
-    private function getContentRenderingConfiguration(): string
-    {
-        return (string)$GLOBALS['TYPO3_CONF_VARS']['FE']['defaultTypoScript_setup.']['defaultContentRendering'];
-    }
-
-    /**
-     * Extracts the class name from something like '...->foo'.
-     *
-     * @param string $reference
-     *
-     * @return class-string
-     */
-    private function extractClassNameFromUserFunction(string $reference): string
-    {
-        $parts = \explode('->', $reference);
-
-        return \array_shift($parts);
-    }
-
-    /**
-     * Extracts the method name from something like '...->foo'.
-     *
-     * @param string $reference
-     *
-     * @return string method name
-     */
-    private function extractMethodNameFromUserFunction(string $reference): string
-    {
-        $parts = \explode('->', $reference);
-
-        return \array_pop($parts);
-    }
-
-    /**
-     * @test
-     */
-    public function defaultContentRenderingIsGenerated(): void
-    {
-        $configuration = $this->getContentRenderingConfiguration();
-
-        self::assertStringContainsString('TypoScript added by extension "seminars"', $configuration);
-        self::assertStringContainsString('tt_content.list.20.seminars_pi1 = < plugin.tx_seminars_pi1', $configuration);
-    }
-
-    /**
-     * @test
-     */
-    public function pluginUserFuncPointsToExistingMethodInExistingDefaultControllerClass(): void
-    {
-        $configuration = $this->getContentRenderingConfiguration();
-
-        $matches = [];
-        \preg_match('/plugin\\.tx_seminars_pi1\\.userFunc = ([^\\s]+)/', $configuration, $matches);
-        $className = $this->extractClassNameFromUserFunction($matches[1]);
-        $methodName = $this->extractMethodNameFromUserFunction($matches[1]);
-
-        self::assertSame(DefaultController::class, $className);
-
-        self::assertTrue(
-            \method_exists(DefaultController::class, $methodName),
-            'Method ' . $methodName . ' does not exist in class ' . $className
-        );
     }
 
     // Tests concerning the list view
